refactor(computeacls): remove reference to API version
