--- fastly/computeacl/v1/doc.go
@@ -1,3 +1,0 @@
-// Package v1 contains functionality which offer various operations to configure
-// Fastly compute ACLs using its v1 implementation.
-package v1
