+++ fastly/computeacls/api_delete.go
@@ -1,4 +1,4 @@
-package v1
+package computeacls
 
 import (
 	"net/http"
