+++ fastly/computeacls/api_list_acls.go
@@ -1,4 +1,4 @@
-package v1
+package computeacls
 
 import (
 	"encoding/json"
