+++ fastly/computeacls/api_lookup.go
@@ -1,4 +1,4 @@
-package v1
+package computeacls
 
 import (
 	"encoding/json"
