+++ fastly/computeacls/api_test.go
@@ -1,4 +1,4 @@
-package v1
+package computeacls
 
 import (
 	"net"
