+++ fastly/computeacls/fixtures/list_entries.yaml
